+++ packages/cli/src/commands/comment/index.ts
@@ -3,6 +3,7 @@ import { defineCommand } from 'citty'
 import add from './add.ts'
 import del from './delete.ts'
 import list from './list.ts'
+import resolve from './resolve.ts'
 import watch from './watch.ts'
 
 export default defineCommand({
@@ -11,6 +12,7 @@ export default defineCommand({
     list,
     add,
     delete: del,
+    resolve,
     watch
   }
 })